requirements.txt
pyaudio==0.2.13
numpy==1.24.0
vosk==0.3.45
requests==2.31.0
pyttsx3==2.90
pyserial==3.5
python-dotenv==1.0.0
pyyaml==6.0
ollama==0.0.19